@@ -12,8 +12,9 @@
         "doctrine/orm": "^2.17",
         "phpdocumentor/reflection-docblock": "^5.3",
         "phpstan/phpdoc-parser": "^1.25",
+        "symfony/apache-pack": "^1.0",
         "symfony/asset": "6.1.*",
-        "symfony/asset-mapper": "6.1.*",
+        "symfony/asset-mapper": "^6.3",
         "symfony/console": "6.1.*",
         "symfony/doctrine-messenger": "6.1.*",
         "symfony/dotenv": "6.1.*",
@@ -102,7 +103,7 @@
     },
     "extra": {
         "symfony": {
-            "allow-contrib": false,
+            "allow-contrib": "true",
             "require": "6.1.*"
         }
     }
